+++ server/src/main/java/org/opensearch/index/IndexSortConfig.java
@@ -143,6 +143,7 @@ public final class IndexSortConfig {
 
     // visible for tests
     final FieldSortSpec[] sortSpecs;
+    final boolean shouldWidenIndexSortType;
 
     public IndexSortConfig(IndexSettings indexSettings) {
         final Settings settings = indexSettings.getSettings();
@@ -182,6 +183,7 @@ public final class IndexSortConfig {
                 sortSpecs[i].missingValue = missingValues.get(i);
             }
         }
+        this.shouldWidenIndexSortType = indexSettings.shouldWidenIndexSortType();
     }
 
     /**
@@ -230,7 +232,11 @@ public final class IndexSortConfig {
             if (fieldData == null) {
                 throw new IllegalArgumentException("docvalues not found for index sort field:[" + sortSpec.field + "]");
             }
-            sortFields[i] = fieldData.sortField(sortSpec.missingValue, mode, null, reverse);
+            if (this.shouldWidenIndexSortType == true) {
+                sortFields[i] = fieldData.wideSortField(sortSpec.missingValue, mode, null, reverse);
+            } else {
+                sortFields[i] = fieldData.sortField(sortSpec.missingValue, mode, null, reverse);
+            }
             validateIndexSortField(sortFields[i]);
         }
         return new Sort(sortFields);